+++ components/site-header/site-header.php
@@ -2,7 +2,10 @@
     <?php
     $name = get_bloginfo('name');
     $logo = wp_get_attachment_image_url(get_option('options_logo'), 'full');
-    $menu = Starterkit_Frontend::get_nav_menu_items_by_location('primary', array('depth' => 2));
+    $menu = [];
+    if(class_exists('Starterkit_Menus')) {
+        $menu = Starterkit_Menus::get_nav_menu_items_by_location('primary', array('depth' => 2));
+    }
     ?>
     <div data-vue-component="site-navigation" xmlns="../vue.xsd">
         <site-navigation